cdk/requirements.txt
aws-cdk.core==1.134.0
aws-cdk.aws-ec2==1.134.0
aws-cdk.aws-s3==1.134.0
